+++ CHANGELOG.md
@@ -2,6 +2,11 @@ CHANGE LOG
 ==========
 
 
+## 5.0.6 (UPCOMING)
+
+* Fixed `ProjectResource` links hydration
+
+
 ## 5.0.5 (03/05/2025)
 
 * Fixed `FirewallRuleOutbound` hydration